Implement authentication and authorization
Authentication is mTLS.
Authorization is based in the CN of the certificate. The methods _Enrol_
and _Registration_ can be called only with a registration certificate
(i.e. cn = "registration.*). The certificate can be expired or revoked
but, as long as mTLS is resolved, it will allow the device to be
registered.

For all other methods, a valid and non revoked certificate is needed. Generally, the certificate cn is
"device_id.domain".
The device_id is sent as metadata by client. Without this metadata, the
authorization fails.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>